requirements.txt
streamlit>=1.36
openai>=1.40
python-dotenv>=1.0
tectonic>=0.1.0
